+++ src/pages/log-page.tsx
@@ -4,13 +4,14 @@ import { Suspense, lazy} from "react";
 const LogContent = lazy(()=>import("@/contents/log"))
 
 interface Props{
-     returnUrl: string
+     returnUrl: string,
+     category?: string
 }
-export default function LogPage({returnUrl}: Props){
+export default function LogPage({returnUrl, category}: Props){
      return (
           <AppLayout className="space-y-4 p-4">
                <Suspense fallback={<LogLoader/>}>
-                    <LogContent returnUrl={returnUrl}/>
+                    <LogContent returnUrl={returnUrl} categoryParam={category}/>
                </Suspense>
           </AppLayout>
      )